test: collect the Windows console test on Windows alone
The hidden-console test launches a Windows process tree and observes the consoles
its descendants attach to, so no other platform can attempt it. Guarding it with
`it.runIf` left every Linux run reporting a skip, and a skip in a summary reads as
a test somebody disabled rather than one that does not apply.

Move it to its own file and leave that file uncollected off Windows. The pure
functions describing Windows conventions stay where they were, because every
platform can and should run those. Windows now reports 953 passed with nothing
skipped, and Linux reports only the tests that apply to it.
