+++ vitest.config.ts
@@ -6,7 +6,14 @@ export default defineConfig({
     // orchestration/worktrees, tests included. Without excluding them the suite collects
     // its own copies out of every task in flight and fails on their unfinished state —
     // which is what happened the first time the core improved itself.
-    exclude: ['**/node_modules/**', '**/dist/**', 'orchestration/worktrees/**'],
+    exclude: [
+      '**/node_modules/**', '**/dist/**', 'orchestration/worktrees/**',
+      // Launching a Windows process tree and observing its consoles is not something
+      // another platform can attempt. Leaving the file uncollected keeps a Linux run
+      // reporting only tests that apply to it; collecting it and reporting a skip reads
+      // as a test somebody disabled, which is the question it kept prompting.
+      ...(process.platform === 'win32' ? [] : ['tests/windows-console.test.ts']),
+    ],
     // Fork IPC times out while the git-heavy suites run on Windows under Node 24;
     // worker threads use the same isolation without that process-channel failure.
     pool: process.platform === 'win32' ? 'threads' : 'forks',
